Separate TypeDoc Transforme class from main module
